reorganize test utils
